requirements.txt
frida-tools>=12.0.0
